+++ DeepLearning/train/training_data.h
@@ -3,8 +3,8 @@
 
 struct training_data{
   size_t input_number;
-  float* inputs;
-  float* expected_output;
+  double* inputs;
+  double* expected_output;
 };
 
 struct training_set{
@@ -20,7 +20,7 @@ struct minibatch_set{
 };
 
 // creates training set based on inputs and outputs for each training data 
-struct training_set create_training_set(float** inputs, float** outputs, size_t data_number, size_t input_number);
+struct training_set create_training_set(double** inputs, double** outputs, size_t data_number, size_t input_number);
 void free_training_set(struct training_set set);
 void print_training_set(struct training_set set);
 void shuffle_training_set(struct training_set set);
