+++ DeepLearning/train/training_data_loader.h
@@ -2,4 +2,4 @@
 #include <stdio.h>
 #include <stddef.h>
 
-void load_training_data(FILE *file, float** *inputs_p, float** *outputs_p, size_t data_number, size_t input_size);
+void load_training_data(FILE *file, double** *inputs_p, double** *outputs_p, size_t data_number, size_t input_size);
